Add shared/ contracts module for run API and stream events
Client API types were manually duplicated from the server, with notable
drift: Activity.vote.result was string on the client where the server
emits "set" | "cleared" | "switched", and the run-stream event names +
payload shapes lived in two stringly-typed switch statements.

shared/contracts.ts is now the single source of truth for everything
that crosses the wire — Post, CommentNode, Snapshot, Participant,
AgentResult, Totals, SimulationResult, RunRecord, ActivityEvent,
SimulationMode, plus the RunStreamEventMap that pins each event name
to its payload. Pure TypeScript, no runtime deps, so importing it
from the client doesn't pull zod into the bundle.

The server's send() in /api/run-stream now takes
<K extends RunStreamEventName>(name: K, data: RunStreamEventMap[K]) so
adding or renaming an event without updating the contract is a compile
error. The client parser uses the same map plus isRunStreamEventName()
to filter unknown events, and the existing useRunSimulation switch is
exhaustively typed against the union.

The zod request schema moved to server/src/runRequestSchema.ts (renamed
from encryptedKey.ts) since it's server-only and Node module resolution
from /shared/ couldn't find /server/node_modules/zod. Test imports were
updated accordingly.

Closes #7
