+++ client/tsconfig.json
@@ -10,7 +10,8 @@
     "skipLibCheck": true,
     "resolveJsonModule": true,
     "types": ["vite/client"],
+    "rootDir": "..",
     "noEmit": true
   },
-  "include": ["src/**/*"]
+  "include": ["src/**/*", "../shared/**/*"]
 }
